refactor(reconciler): keep the config-merge rules in one file
affinity.go was the wrong unit. This package splits by concern — apply
semantics, cleanup, health, events, discovery — and "one CRD field's decode and
merge" is finer than any of them.

It also separated two siblings that do the same job at the same layer, so
reading MergeRoleGroupConfig meant jumping files for one of its two helpers and
scrolling twenty lines for the other:

    merged.Affinity  = mergeAffinity(...)     // affinity.go
    merged.Resources = mergeResources(...)    // right here

The affinity helpers now sit immediately after mergeResources, where the rest of
the merge rules already live. DecodeAffinity goes with them: its caller is in
base_role_group_handler.go, but that is the existing relationship between these
two files — role_group_handler.go holds the shared types and helpers
(RoleGroupBuildContext, MergeRoleGroupConfig, RenderLoggingConfigMapData) and
base_role_group_handler.go consumes them.

role_group_handler.go goes from 467 to 636 lines, in line with the package's
other concern files. affinity_internal_test.go is renamed to match its subject;
affinity_test.go keeps its name, since topic-scoped test files are ordinary.

No behaviour change.
